moved logout function to LogoutButton component instead of passing as props
